requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
scipy>=1.11
networkx>=3.2
matplotlib>=3.8
joblib>=1.3
torch>=2.2
sentence-transformers>=2.6
PyYAML>=6.0
tqdm>=4.66
